+++ R/alm.R
@@ -145,6 +145,8 @@ alm.gp <- function(object, x_cand, batch_size = 1, workers = 1, ...) {
       idx_x_cand <- idx_x_cand0[-idx]
     }
   }
+  pkg.env$py_gc$collect()
+  gc(full=T)
   return(idx)
 }
 
@@ -250,6 +252,8 @@ alm.dgp <- function(object, x_cand, batch_size = 1, workers = 1, threading = FAL
     }
     idx <- matrix(idx, nrow = batch_size, byrow = T)
   }
+  pkg.env$py_gc$collect()
+  gc(full=T)
   return(idx)
 }
 
@@ -385,5 +389,7 @@ alm.bundle <- function(object, x_cand, batch_size = 1, workers = 1, threading = 
     }
     idx <- matrix(idx, nrow = batch_size, byrow = T)
   }
+  pkg.env$py_gc$collect()
+  gc(full=T)
   return(idx)
 }